+++ src/fastapi_factory_utilities/core/plugins/odm_plugin/configs.py
@@ -13,3 +13,11 @@ class ODMConfig(BaseModel):
     database: str = "test"
 
     connection_timeout_ms: int = 4000
+
+    min_pool_size: int = 0
+
+    max_pool_size: int = 100
+
+    max_idle_time_ms: int | None = None
+
+    heartbeat_frequency_ms: int | None = None